Move shape-shape intersection and casting to the Shape class.
