+++ src.ts/geometry/shape.ts
@@ -1,14 +1,114 @@
-import {Vector, VectorOps} from "../math"
-import {RawShape} from "../raw";
+import { Vector, VectorOps, Rotation, RotationOps } from "../math"
+import { RawShape } from "../raw";
+import { ShapeTOI } from "./toi";
+
+export class Shape {
+    public intoRaw?(): RawShape;
+
+    /**
+     * Computes the time of impact between two moving shapes.
+     * @param shapePos1 - The initial position of this sahpe.
+     * @param shapeRot1 - The rotation of this shape.
+     * @param shapeVel1 - The velocity of this shape.
+     * @param shape2 - The second moving shape.
+     * @param shapePos2 - The initial position of the second shape.
+     * @param shapeRot2 - The rotation of the second shape.
+     * @param shapeVel2 - The velocity of the second shape.
+     * @param maxToi - The maximum time when the inpact can happen.
+     * @returns If the two moving shapes collider at some point along their trajectories, this retruns the
+     *  time at which the two shape collider as well as the contact information durning the impact. Returns
+     *  `null`if the two shapes never collide along their paths.
+     */
+    public castShape(
+        shapePos1: Vector,
+        shapeRot1: Rotation,
+        shapeVel1: Vector,
+        shape2: Shape,
+        shapePos2: Vector,
+        shapeRot2: Rotation,
+        shapeVel2: Vector,
+        maxToi: number
+    ): ShapeTOI | null {
+        let rawPos1 = VectorOps.intoRaw(shapePos1);
+        let rawRot1 = RotationOps.intoRaw(shapeRot1);
+        let rawVel1 = VectorOps.intoRaw(shapeVel1);
+        let rawPos2 = VectorOps.intoRaw(shapePos2);
+        let rawRot2 = RotationOps.intoRaw(shapeRot2);
+        let rawVel2 = VectorOps.intoRaw(shapeVel2);
+
+        let rawShape1 = this.intoRaw();
+        let rawShape2 = shape2.intoRaw();
+
+        let result = ShapeTOI.fromRaw(rawShape1.castShape(
+            rawPos1,
+            rawRot1,
+            rawVel1,
+            rawShape2,
+            rawPos2,
+            rawRot2,
+            rawVel2,
+            maxToi
+        ));
+
+        rawPos1.free();
+        rawRot1.free();
+        rawVel1.free();
+        rawPos2.free();
+        rawRot2.free();
+        rawVel2.free();
+
+        rawShape1.free();
+        rawShape2.free();
+
+        return result;
+    }
+
+    /**
+     * Tests if this shape intersects another shape.
+     * 
+     * @param shapePos1 - The position of this shape.
+     * @param shapeRot1 - The rotation of this shape.
+     * @param shape2  - The second shape to test.
+     * @param shapePos2 - The position of the second shape.
+     * @param shapeRot2 - The rotation of the second shape.
+     * @returns `true` if the two shapes intersect, `false` if they don’t.
+     */
+    public intersectsShape(
+        shapePos1: Vector,
+        shapeRot1: Rotation,
+        shape2: Shape,
+        shapePos2: Vector,
+        shapeRot2: Rotation
+    ): boolean {
+        let rawPos1 = VectorOps.intoRaw(shapePos1);
+        let rawRot1 = RotationOps.intoRaw(shapeRot1);
+        let rawPos2 = VectorOps.intoRaw(shapePos2);
+        let rawRot2 = RotationOps.intoRaw(shapeRot2);
+
+        let rawShape1 = this.intoRaw();
+        let rawShape2 = shape2.intoRaw();
+
+        let result = rawShape1.intersectsShape(
+            rawPos1,
+            rawRot1,
+            rawShape2,
+            rawPos2,
+            rawRot2
+        );
+
+        rawPos1.free();
+        rawRot1.free();
+        rawPos2.free();
+        rawRot2.free();
+
+        rawShape1.free();
+        rawShape2.free();
+
+        return result;
+    }
+}
 
 // #if DIM2
-/**
- * The type of a shape supported by Rapier.
- */
-export type Shape = Ball | Cuboid | Capsule | Segment | Triangle
-    | TriMesh | Heightfield | ConvexPolygon | RoundCuboid
-    | RoundTriangle | RoundConvexPolygon;
-
 /**
  * An enumeration representing the type of a shape.
  */
@@ -31,13 +131,6 @@ export enum ShapeType {
 // #endif
 
 // #if DIM3
-/**
- * The type of a shape supported by Rapier.
- */
-export type Shape = Ball | Cuboid | Capsule | Segment | Triangle
-    | TriMesh | Heightfield | ConvexPolyhedron | Cylinder
-    | Cone | RoundCuboid | RoundCylinder | RoundCone | RoundConvexPolyhedron
-
 /**
  * An enumeration representing the type of a shape.
  */
@@ -66,7 +159,7 @@ export enum ShapeType {
 /**
  * A shape that is a sphere in 3D and a circle in 2D.
  */
-export class Ball {
+export class Ball extends Shape {
     /**
      * The balls radius.
      */
@@ -77,6 +170,7 @@ export class Ball {
      * @param radius - The balls radius.
      */
     constructor(radius: number) {
+        super();
         this.radius = radius;
     }
 
@@ -88,7 +182,7 @@ export class Ball {
 /**
  * A shape that is a box in 3D and a rectangle in 2D.
  */
-export class Cuboid {
+export class Cuboid extends Shape {
     /**
      * The half extent of the cuboid along each coordinate axis.
      */
@@ -101,6 +195,7 @@ export class Cuboid {
      * @param hy - The helf height of the rectangle.
      */
     constructor(hx: number, hy: number) {
+        super();
         this.halfExtents = VectorOps.new(hx, hy);
     }
 
@@ -114,6 +209,7 @@ export class Cuboid {
      * @param hz - The half depth of the cuboid.
      */
     constructor(hx: number, hy: number, hz: number) {
+        super();
         this.halfExtents = VectorOps.new(hx, hy, hz);
     }
 
@@ -134,7 +230,7 @@ export class Cuboid {
 /**
  * A shape that is a box in 3D and a rectangle in 2D, with round corners.
  */
-export class RoundCuboid {
+export class RoundCuboid extends Shape {
     /**
      * The half extent of the cuboid along each coordinate axis.
      */
@@ -154,6 +250,7 @@ export class RoundCuboid {
      *   effectively increase the half-extents of the cuboid by this radius.
      */
     constructor(hx: number, hy: number, borderRadius: number) {
+        super();
         this.halfExtents = VectorOps.new(hx, hy);
         this.borderRadius = borderRadius;
     }
@@ -170,6 +267,7 @@ export class RoundCuboid {
      *   effectively increase the half-extents of the cuboid by this radius.
      */
     constructor(hx: number, hy: number, hz: number, borderRadius: number) {
+        super();
         this.halfExtents = VectorOps.new(hx, hy, hz);
         this.borderRadius = borderRadius;
     }
@@ -190,7 +288,7 @@ export class RoundCuboid {
 /**
  * A shape that is a capsule.
  */
-export class Capsule {
+export class Capsule extends Shape {
     /**
      * The radius of the capsule's basis.
      */
@@ -207,6 +305,7 @@ export class Capsule {
      * @param radius - The balls radius.
      */
     constructor(halfHeight: number, radius: number) {
+        super();
         this.halfHeight = halfHeight;
         this.radius = radius;
     }
@@ -219,7 +318,7 @@ export class Capsule {
 /**
  * A shape that is a segment.
  */
-export class Segment {
+export class Segment extends Shape {
     /**
      * The first point of the segment.
      */
@@ -236,6 +335,7 @@ export class Segment {
      * @param b - The second point of the segment.
      */
     constructor(a: Vector, b: Vector) {
+        super();
         this.a = a;
         this.b = b;
     }
@@ -253,7 +353,7 @@ export class Segment {
 /**
  * A shape that is a segment.
  */
-export class Triangle {
+export class Triangle extends Shape {
     /**
      * The first point of the triangle.
      */
@@ -277,6 +377,7 @@ export class Triangle {
      * @param c - The third point of the triangle.
      */
     constructor(a: Vector, b: Vector, c: Vector) {
+        super();
         this.a = a;
         this.b = b;
         this.c = c;
@@ -298,7 +399,7 @@ export class Triangle {
 /**
  * A shape that is a triangle with round borders and a non-zero thickness.
  */
-export class RoundTriangle {
+export class RoundTriangle extends Shape {
     /**
      * The first point of the triangle.
      */
@@ -330,6 +431,7 @@ export class RoundTriangle {
      *   this is also equal to half the thickness of the triangle.
      */
     constructor(a: Vector, b: Vector, c: Vector, borderRadius: number) {
+        super();
         this.a = a;
         this.b = b;
         this.c = c;
@@ -351,7 +453,7 @@ export class RoundTriangle {
 /**
  * A shape that is a triangle mesh.
  */
-export class Polyline {
+export class Polyline extends Shape {
     /**
      * The vertices of the polyline.
      */
@@ -370,6 +472,7 @@ export class Polyline {
      *    the vertices are assumed to form a line strip.
      */
     constructor(vertices: Float32Array, indices: Uint32Array) {
+        super();
         this.vertices = vertices;
         this.indices = !!indices ? indices : new Uint32Array(0);
     }
@@ -382,7 +485,7 @@ export class Polyline {
 /**
  * A shape that is a triangle mesh.
  */
-export class TriMesh {
+export class TriMesh extends Shape {
     /**
      * The vertices of the triangle mesh.
      */
@@ -400,6 +503,7 @@ export class TriMesh {
      * @param indices - The indices of the triangle mesh's triangles.
      */
     constructor(vertices: Float32Array, indices: Uint32Array) {
+        super();
         this.vertices = vertices;
         this.indices = indices;
     }
@@ -414,7 +518,7 @@ export class TriMesh {
 /**
  * A shape that is a convex polygon.
  */
-export class ConvexPolygon {
+export class ConvexPolygon extends Shape {
     /**
      * The vertices of the convex polygon.
      */
@@ -434,6 +538,7 @@ export class ConvexPolygon {
      *   be done automatically.
      */
     constructor(vertices: Float32Array, skipConvexHullComputation: boolean) {
+        super();
         this.vertices = vertices;
         this.skipConvexHullComputation = !!skipConvexHullComputation;
     }
@@ -450,7 +555,7 @@ export class ConvexPolygon {
 /**
  * A shape that is a convex polygon.
  */
-export class RoundConvexPolygon {
+export class RoundConvexPolygon extends Shape {
     /**
      * The vertices of the convex polygon.
      */
@@ -476,6 +581,7 @@ export class RoundConvexPolygon {
      *   be done automatically.
      */
     constructor(vertices: Float32Array, borderRadius: number, skipConvexHullComputation: boolean) {
+        super();
         this.vertices = vertices;
         this.borderRadius = borderRadius;
         this.skipConvexHullComputation = !!skipConvexHullComputation;
@@ -493,7 +599,7 @@ export class RoundConvexPolygon {
 /**
  * A shape that is a heightfield.
  */
-export class Heightfield {
+export class Heightfield extends Shape {
     /**
      * The heights of the heightfield, along its local `y` axis.
      */
@@ -511,6 +617,7 @@ export class Heightfield {
      * @param scale - The scale factor applied to the heightfield.
      */
     constructor(heights: Float32Array, scale: Vector) {
+        super();
         this.heights = heights;
         this.scale = scale;
     }
@@ -530,7 +637,7 @@ export class Heightfield {
 /**
  * A shape that is a convex polygon.
  */
-export class ConvexPolyhedron {
+export class ConvexPolyhedron extends Shape {
     /**
      * The vertices of the convex polygon.
      */
@@ -552,6 +659,7 @@ export class ConvexPolyhedron {
      *   is already convex.
      */
     constructor(vertices: Float32Array, indices: Uint32Array | null) {
+        super();
         this.vertices = vertices;
         this.indices = indices;
     }
@@ -569,7 +677,7 @@ export class ConvexPolyhedron {
 /**
  * A shape that is a convex polygon.
  */
-export class RoundConvexPolyhedron {
+export class RoundConvexPolyhedron extends Shape {
     /**
      * The vertices of the convex polygon.
      */
@@ -596,6 +704,7 @@ export class RoundConvexPolyhedron {
      * @param borderRadius - The radius of the borders of this convex polyhedron.
      */
     constructor(vertices: Float32Array, indices: Uint32Array | null, borderRadius: number) {
+        super();
         this.vertices = vertices;
         this.indices = indices;
         this.borderRadius = borderRadius;
@@ -613,7 +722,7 @@ export class RoundConvexPolyhedron {
 /**
  * A shape that is a heightfield.
  */
-export class Heightfield {
+export class Heightfield extends Shape {
     /**
      * The number of rows in the heights matrix.
      */
@@ -645,6 +754,7 @@ export class Heightfield {
      * @param scale - The dimensions of the heightfield's local `x,z` plane.
      */
     constructor(nrows: number, ncols: number, heights: Float32Array, scale: Vector) {
+        super();
         this.nrows = nrows;
         this.ncols = ncols;
         this.heights = heights;
@@ -662,7 +772,7 @@ export class Heightfield {
 /**
  * A shape that is a 3D cylinder.
  */
-export class Cylinder {
+export class Cylinder extends Shape {
     /**
      * The radius of the cylinder's basis.
      */
@@ -679,6 +789,7 @@ export class Cylinder {
      * @param radius - The balls radius.
      */
     constructor(halfHeight: number, radius: number) {
+        super();
         this.halfHeight = halfHeight;
         this.radius = radius;
     }
@@ -692,7 +803,7 @@ export class Cylinder {
 /**
  * A shape that is a 3D cylinder with round corners.
  */
-export class RoundCylinder {
+export class RoundCylinder extends Shape {
     /**
      * The radius of the cylinder's basis.
      */
@@ -715,6 +826,7 @@ export class RoundCylinder {
      * @param borderRadius - The radius of the borders of this cylinder.
      */
     constructor(halfHeight: number, radius: number, borderRadius: number) {
+        super();
         this.borderRadius = borderRadius;
         this.halfHeight = halfHeight;
         this.radius = radius;
@@ -728,7 +840,7 @@ export class RoundCylinder {
 /**
  * A shape that is a 3D cone.
  */
-export class Cone {
+export class Cone extends Shape {
     /**
      * The radius of the cone's basis.
      */
@@ -745,6 +857,7 @@ export class Cone {
      * @param radius - The balls radius.
      */
     constructor(halfHeight: number, radius: number) {
+        super();
         this.halfHeight = halfHeight;
         this.radius = radius;
     }
@@ -757,7 +870,7 @@ export class Cone {
 /**
  * A shape that is a 3D cone with round corners.
  */
-export class RoundCone {
+export class RoundCone extends Shape {
     /**
      * The radius of the cone's basis.
      */
@@ -780,6 +893,7 @@ export class RoundCone {
      * @param borderRadius - The radius of the borders of this cone.
      */
     constructor(halfHeight: number, radius: number, borderRadius: number) {
+        super();
         this.halfHeight = halfHeight;
         this.radius = radius;
         this.borderRadius = borderRadius;
